+++ CHANGES.md
@@ -11,6 +11,13 @@
   * Rendering should be configurable (default: fn that returns file-content as body)
 
 
+## 0.4.0 / 2016-June-??
+
+* Helper fn to build routes from given route specs
+  * [TODO] Support for splitting routes with URI and method specifiers into nested routes
+* [TODO] A truthy `(true? matcher)` should result in direct dispatch
+
+
 ## 0.3.0 / 2016-May-27
 
 * Support for extensible routes as a first-class abstraction